Refactoring dataAnalyzer.py. Splitting into 4 scripts: database_metadata.py, utils.py, clustering.py, classification.py. Now they can be used as separate programs.
