+++ utils.py
@@ -1,0 +1,650 @@
+"""
+    Predicting trajectories of objects
+    Copyright (C) 2022  Bence Peter
+
+    This program is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+    Contact email: ecneb2000@gmail.com
+"""
+import matplotlib.pyplot as plt
+from dataManagementClasses import Detection, TrackedObject
+import numpy as np
+import time
+import databaseLoader
+import tqdm
+import os
+import joblib 
+
+def savePlot(fig: plt.Figure, name: str):
+    fig.savefig(name, dpi=150)
+
+def detectionFactory(objID: int, frameNum: int, label: str, confidence: float, x: float, y: float, width: float, height: float, vx: float, vy: float, ax:float, ay: float) -> Detection:
+    """Create Detection object.
+
+    Args:
+        objID (int): Object ID, to which object the Detection belongs to. 
+        frameNum (int): Frame number when the Detection occured. 
+        label (str): Label of the object, etc: car, person... 
+        confidence (float): Confidence number, of how confident the neural network is in the detection. 
+        x (float):  X coordinate of the object.
+        y (float): Y coordinate of the object. 
+        width (float): Width of the bounding box of the object. 
+        height (float): Height of the bounging box of the object. 
+        vx (float): Velocity on the X axis. 
+        vy (float): Velocity on the Y axis. 
+        ax (float): Accelaration on the X axis. 
+        ay (float): Accelaration on the Y axis. 
+
+    Returns:
+        Detection: The Detection object, which is to be returned. 
+    """
+    retDet = Detection(label, confidence, x,y,width,height,frameNum)
+    retDet.objID = objID
+    retDet.VX = vx
+    retDet.VY = vy
+    retDet.AX = ax
+    retDet.AY = ay
+    return retDet
+
+def trackedObjectFactory(detections: list) -> TrackedObject:
+    """Create trackedObject object from list of detections
+
+    Args:
+        detections (list): list of detection 
+
+    Returns:
+        TrackedObject:  trackedObject
+    """
+    tmpObj = TrackedObject(detections[0].objID, detections[0], len(detections))
+    for det in detections:
+        tmpObj.history.append(det)
+        tmpObj.X = det.X
+        tmpObj.Y = det.Y
+        tmpObj.VX = det.VX
+        tmpObj.VY = det.VY
+        tmpObj.AX = det.AX
+        tmpObj.AY = det.AY
+    return tmpObj
+
+def cvCoord2npCoord(Y: np.ndarray) -> np.ndarray:
+    """Convert OpenCV Y axis coordinates to numpy coordinates.
+
+    Args:
+        Y (np.ndarray): Y axis coordinate vector
+
+    Returns:
+        np.ndarray: Y axis coordinate vector
+    """
+    return 1 - Y
+
+def detectionParser(rawDetectionData) -> list:
+    """Parse raw detection data loaded from database.
+    Returns a list of detections. 
+
+    Args:
+        rawDetectionData (list): list of raw detection entries
+
+    Returns:
+        detections: list of parsed detections 
+    """
+    detections = []
+    for entry in rawDetectionData:
+        detections.append(detectionFactory(entry[0], entry[1], entry[2], entry[3], entry[4], entry[5], entry[6], entry[7], entry[8], entry[9], entry[10], entry[11]))
+    return detections 
+
+def parseRawObject2TrackedObject(rawObjID: int, path2db: str):
+    """Takes an objID and the path 2 database, then returns a trackedObject object if detections can be assigned to the object.
+
+    Args:
+        rawObjID (int): ID of an object 
+        path2db (str): path to database 
+
+    Returns:
+        trackedObject: trackedObject object from dataManagement class, if no dets can be assigned to it, then returns False 
+    """
+    rawDets = databaseLoader.loadDetectionsOfObject(path2db, rawObjID)
+    if len(rawDets) > 0:
+        return trackedObjectFactory(detectionParser(rawDets))
+    else:
+        return False
+
+def preprocess_database_data(path2db: str):
+    """Preprocessing database data (detections). Assigning detections to objects.
+
+    Args:
+        path2db (str): Path to database file. 
+
+    Returns:
+        list: list of object tracks 
+    """
+    rawObjectData = databaseLoader.loadObjects(path2db)
+    trackedObjects = []
+    for rawObj in tqdm.tqdm(rawObjectData, desc="Loading detections of tracks."):
+        tmpDets = []
+        rawDets = databaseLoader.loadDetectionsOfObject(path2db, rawObj[0])
+        if len(rawDets) > 0:
+            tmpDets = detectionParser(rawDets)
+            trackedObjects.append(trackedObjectFactory(tmpDets))
+    return trackedObjects
+
+def preprocess_database_data_multiprocessed(path2db: str, n_jobs=None):
+    """Preprocessing database data (detections). Assigning detections to objects.
+    This is the multoprocessed variant of the preprocess_database_data() func.
+
+    Args:
+        path2db (str): Path to database file. 
+
+    Returns:
+        list: list of object tracks 
+    """
+    from multiprocessing import Pool
+    rawObjectData = databaseLoader.loadObjects(path2db)
+    tracks = []
+    with Pool(processes=n_jobs) as pool:
+        print("Preprocessing started.")
+        start = time.time()
+        results = pool.starmap_async(parseRawObject2TrackedObject, [[rawObj[0], path2db] for rawObj in rawObjectData])
+        for result in tqdm.tqdm(results.get(), desc="Unpacking the result of detection assignment."):
+            if result:
+                tracks.append(result)
+        print(f"Detections assigned to Objects in {time.time()-start}s")
+    return tracks
+ 
+def makeColormap(path2db):
+    """Make colormap based on number of objects logged in the database.
+    This colormap vector will be input to matplotlib scatter plot.
+
+    Args:
+        path2db (str): Path to database 
+
+    Returns:
+        colormap: list of color gradient vectors (R,G,B)
+    """
+    objects = databaseLoader.loadObjects(path2db)
+    detections = databaseLoader.loadDetections(path2db)
+    objectVector = []
+    detectionVector = []
+    [objectVector.append(obj) for obj in objects]
+    [detectionVector.append(det) for det in detections]
+    colors = np.random.rand(len(objectVector))
+    colormap = np.zeros(shape=(len(detectionVector)))
+    for i in range(len(objectVector)):
+        for j in range(len(detectionVector)):
+            if objectVector[i][0] == detectionVector[j][0]:
+                colormap[j] = colors[i]
+    return colormap
+
+def findEnterAndExitPoints(path2db: str):
+    """Extracting only the first and the last detections of tracked objects.
+
+    Args:
+        path2db (str): Path to the database file. 
+
+    Returns:
+        enterDetection, exitDetections: List of first and last detections of objects. 
+    """
+    rawDetectionData = databaseLoader.loadDetections(path2db)
+    detections = detectionParser(rawDetectionData)
+    rawObjectData = databaseLoader.loadObjects(path2db)
+    trackedObjects = []
+    for obj in tqdm.tqdm(rawObjectData, desc="Filter out enter and exit points."):
+        tmpDets = []
+        for det in detections:
+           if det.objID == obj[0]:
+            tmpDets.append(det)
+        if len(tmpDets) > 0:
+            trackedObjects.append(trackedObjectFactory(tmpDets))
+    enterDetections = [obj.history[0] for obj in trackedObjects]
+    exitDetections = [obj.history[-1] for obj in trackedObjects]
+    return enterDetections, exitDetections 
+
+def order2track(args: list):
+    """Orders enter and exit dets to object.
+
+    Args:
+        args (list): list of args, that should look like [detections, objID] 
+
+    Returns:
+        tuple: tuple of dets, enter det and exit det 
+        bool: returns False, when tuple is empty
+    """
+    dets = []
+    for det in args[0]:
+        if det.objID == args[1]:
+            dets.append(det)
+    if dets:
+        return (dets[0], dets[-1])
+    else:
+        return False
+
+def findEnterAndExitPointsMultiprocessed(path2db: str, n_jobs=None):
+    """Extract only the first and last detections of tracked objects.
+    This is a multithreaded implementation of function findEnterAndExitPoints(path2db: str)
+
+    Args:
+        path2db (str): Path to database file. 
+    """
+    from multiprocessing import Pool
+    rawDetectionData = databaseLoader.loadDetections(path2db)
+    detections = detectionParser(rawDetectionData)
+    rawObjectData = databaseLoader.loadObjects(path2db)
+    enterDetections = []
+    exitDetections = []
+    iterable = [[detections, obj[0]] for obj in rawObjectData]
+    with Pool(processes=n_jobs) as executor:
+        start = time.time()
+        results = executor.map(order2track, iterable) 
+        print("Enter and Exit points with multiprocessing: %f"%(time.time()-start))
+        for result in tqdm.tqdm(results, desc="Enter and Exit points."):
+            if result:
+                enterDetections.append(result[0])
+                exitDetections.append(result[1])
+    return enterDetections, exitDetections
+
+def euclidean_distance(q1: float, p1: float, q2: float, p2: float):
+    """Calculate euclidean distance of 2D vectors.
+
+    Args:
+        q1 (float): First element of vector1. 
+        p1 (float): First element of vector2. 
+        q2 (float): Second element of vector1. 
+        p2 (float): Second element of vector2.
+        
+    Returns:
+        float: euclidean distance of the two vectors. 
+    """
+    return (((q1-p1)**2)+((q2-p2)**2))**0.5
+    
+
+def filter_out_false_positive_detections(trackedObjects: list, threshold: float):
+    """This function is to filter out false positive detections, 
+    that enter and exit points are closer, than the threshold value given in the arguments.
+
+    Args:
+        trackedObjects (list): list of object trackings 
+        threshold (float): if enter and exit points are closer than this value, they are exclueded from the return list
+
+    Returns:
+        list: list of returned filtered tracks 
+    """
+    filteredTracks = []
+    for obj in tqdm.tqdm(trackedObjects, desc="False positive filtering."):
+        d = euclidean_distance(obj.history[0].X, obj.history[-1].X, obj.history[0].Y, obj.history[-1].Y)
+        if d > threshold:
+            filteredTracks.append(obj)
+    return filteredTracks
+
+def filter_out_edge_detections(trackedObjects: list, threshold: float):
+    """Filter out objects, that enter and exit detections coordinates are in the threshold value.
+
+    Args:
+        trackedObjects (list): list of object trackings 
+        threshold (float): objects only under this value will be returned 
+
+    Returns:
+        list: filtered list of tracks 
+    """
+    max_y = 0 
+    min_y = 9999 
+    max_x = 0
+    min_x = 9999
+    for obj in tqdm.tqdm(trackedObjects, desc="Looking for min max values."):
+        local_min_x = np.min([det.X for det in obj.history])
+        local_max_x = np.max([det.X for det in obj.history])
+        local_min_y = np.min([det.Y for det in obj.history])
+        local_max_y = np.max([det.Y for det in obj.history])
+        if max_x < local_max_x:
+            max_x = local_max_x
+        if min_x > local_min_x:
+            min_x = local_min_x
+        if max_y < local_max_y:
+            max_y = local_max_y
+        if min_y > local_min_y:
+            min_y = local_min_y 
+    #print(f"\n Max X: {max_x}\n Min X: {min_x}\n Max Y: {max_y}\n Min Y: {min_y}\n")
+    #print(f"\n Thresholds:\n Max X: {max_x-threshold}\n Min X: {min_x+threshold}\n Max Y: {max_y-threshold}\n Min Y: {min_y+threshold}\n")
+    filteredTracks = []
+    for obj in tqdm.tqdm(trackedObjects, desc="Filter out edge detections."):
+        if (((obj.history[0].X <= min_x+threshold or obj.history[0].X >= max_x-threshold) or 
+            (obj.history[0].Y <= min_y+threshold or obj.history[0].Y >= max_y-threshold)) and
+            ((obj.history[-1].X <= min_x+threshold or obj.history[-1].X >= max_x-threshold) or 
+            (obj.history[-1].Y <= min_y+threshold or obj.history[-1].Y >= max_y-threshold))): 
+            filteredTracks.append(obj)
+    # even though we did the edge filtering, we can run an euclidean distance based filtering, which's threshold is hardcoded for now
+    return filter_out_false_positive_detections(filteredTracks, 0.4)
+
+def filter_tracks(trackedObjects: list, label="car"):
+    """Only return objects with given label.
+
+    Args:
+        trackedObjects (list): list of tracked objects 
+        label (str, optional): label of object. Defaults to "car".
+
+    Returns:
+        list: list of "label" objects 
+    """
+    return [obj for obj in trackedObjects if obj.label==label]
+
+
+def makeFeatureVectors_Nx2(trackedObjects: list) -> np.ndarray:
+    """Create 2D feature vectors from tracks.
+    The enter and exit coordinates are put in different vectors. Only creating 2D vectors.
+
+    Args:
+        trackedObjects (list): list of tracked objects 
+
+    Returns:
+        np.ndarray: numpy array of feature vectors 
+    """
+    featureVectors = [] 
+    for obj in trackedObjects:
+        featureVectors.append(obj.history[0].X, obj.history[0].Y)
+        featureVectors.append(obj.history[-1].X, obj.history[-1].Y)
+    return np.array(featureVectors)
+
+def makeFeatureVectorsNx4(trackedObjects: list) -> np.ndarray:
+    """Create 4D feature vectors from tracks.
+    The enter and exit coordinates are put in one vector. Creating 4D vectors.
+    v = [enterX, enterY, exitX, exitY]
+
+    Args:
+        trackedObjects (list): list of tracked objects 
+
+    Returns:
+        np.ndarray: numpy array of feature vectors 
+    """
+    featureVectors = np.array([np.array([obj.history[0].X, obj.history[0].Y, obj.history[-1].X, obj.history[-1].Y]) for obj in tqdm.tqdm(trackedObjects, desc="Feature vectors.")])
+    return featureVectors
+
+def shuffle_data(trackedObjects: list) -> list:
+    rng = np.random.default_rng()
+    for i in range(len(trackedObjects)):
+        randIDX = int(len(trackedObjects) * rng.random())
+        tmpObj = trackedObjects[i]
+        trackedObjects[i] = trackedObjects[randIDX]
+        trackedObjects[randIDX] = tmpObj
+    
+def test_shuffle(path2db: str):
+    origObjects = preprocess_database_data_multiprocessed(path2db=path2db, n_jobs=16)
+    objectsToShuffle = origObjects.copy()
+    shuffle_data(objectsToShuffle)
+    changed = 0
+    for i in range(len(origObjects)):
+        if origObjects[i].objID != objectsToShuffle[i].objID:
+            changed += 1
+    print(f"Entropy = {changed/len(origObjects)}")
+
+
+def checkDir(path2db):
+    """Check for dir of given database, to be able to save plots.
+
+    Args:
+        path2db (str): Path to database. 
+    """
+    if not os.path.isdir(os.path.join("research_data", path2db.split('/')[-1].split('.')[0])):
+        os.mkdir(os.path.join("research_data", path2db.split('/')[-1].split('.')[0]))
+        print("Directory \"research_data/{}\" is created.".format(path2db.split('/')[-1].split('.')[0]))
+
+def make_features_for_classification(trackedObjects: list, k: int, labels: np.ndarray):
+    """Make feature vectors for classification algorithm
+
+    Args:
+        trackedObjects (list): Tracked objects 
+        k (int): K is the number of slices, the object history should be sliced up into.
+        labels (np.ndarray): Results of clustering.
+
+    Returns:
+        np.ndarray: featurevectors and the new labels to fit the featurevectors 
+    """
+    featureVectors = []
+    newLabels = []
+    for j in range(len(trackedObjects)):
+        step = len(trackedObjects[j].history)//k
+        if step > 0:
+            midstep = step//2
+            for i in range(0, len(trackedObjects[j].history)-step, step):
+                featureVectors.append(np.array([trackedObjects[j].history[i].X,trackedObjects[j].history[i].Y,trackedObjects[j].history[i+midstep].X,trackedObjects[j].history[i+midstep].Y,trackedObjects[j].history[i+step].X,trackedObjects[j].history[i+step].Y]))
+                newLabels.append(labels[j])
+    return np.array(featureVectors), np.array(newLabels)
+
+def make_features_for_classification_velocity(trackedObjects: list, k: int, labels: np.ndarray):
+    """Make feature vectors for classification algorithm
+
+    Args:
+        trackedObjects (list): Tracked objects 
+        k (int): K is the number of slices, the object history should be sliced up into.
+        labels (np.ndarray): Results of clustering.
+
+    Returns:
+        np.ndarray, np.ndarray: featureVectors, labels 
+    """
+    featureVectors = []
+    newLabels = []
+    for j in range(len(trackedObjects)):
+        step = len(trackedObjects[j].history)//k
+        if step > 0:
+            midstep = step//2
+            for i in range(0, len(trackedObjects[j].history)-step, step):
+                featureVectors.append(np.array([trackedObjects[j].history[i].X,trackedObjects[j].history[i].Y,trackedObjects[j].history[i].VX,trackedObjects[j].history[i].VY,trackedObjects[j].history[i+midstep].X,trackedObjects[j].history[i+midstep].Y,trackedObjects[j].history[i+step].X,trackedObjects[j].history[i+step].Y,trackedObjects[j].history[i+step].VX,trackedObjects[j].history[i+step].VY]))
+                newLabels.append(labels[j])
+    return np.array(featureVectors), np.array(newLabels)
+
+def make_features_for_classification_velocity_time(trackedObjects: list, k: int, labels: np.ndarray):
+    """Make feature vectors for classification algorithm
+
+    Args:
+        trackedObjects (list): Tracked objects 
+        k (int): K is the number of slices, the object history should be sliced up into.
+        labels (np.ndarray): Results of clustering.
+
+    Returns:
+        np.ndarray, np.ndarray, np.ndarray: featureVectors, labels, timeOfFeatureVectors
+    """
+    featureVectors = []
+    newLabels = []
+    time = []
+    for j in range(len(trackedObjects)):
+        step = len(trackedObjects[j].history)//k
+        if step > 0:
+            midstep = step//2
+            for i in range(0, len(trackedObjects[j].history)-step, step):
+                featureVectors.append(np.array([trackedObjects[j].history[i].X,trackedObjects[j].history[i].Y,trackedObjects[j].history[i].VX,trackedObjects[j].history[i].VY,trackedObjects[j].history[i+midstep].X,trackedObjects[j].history[i+midstep].Y,trackedObjects[j].history[i+step].X,trackedObjects[j].history[i+step].Y,trackedObjects[j].history[i+step].VX,trackedObjects[j].history[i+step].VY]))
+                newLabels.append(labels[j])
+                time.append(np.array([trackedObjects[j].history[i].frameID, trackedObjects[j].history[i+midstep].frameID, trackedObjects[j].history[i+step].frameID]))
+    return np.array(featureVectors), np.array(newLabels), np.array(time)
+
+def make_features_for_classification_velocity_time_second_half(trackedObjects: list, k: int, labels: np.ndarray):
+    """Make feature vectors for classification algorithm
+
+    Args:
+        trackedObjects (list): Tracked objects 
+        k (int): K is the number of slices, the object history should be sliced up into.
+        labels (np.ndarray): Results of clustering.
+
+    Returns:
+        np.ndarray, np.ndarray, np.ndarray: featureVectors, labels, timeOfFeatureVectors
+    """
+    featureVectors = []
+    newLabels = []
+    time = []
+    for j in range(len(trackedObjects)):
+        step = (len(trackedObjects[j].history)//2)//k
+        if step > 0:
+            midstep = step//2
+            for i in range(len(trackedObjects[j].history)//2, len(trackedObjects[j].history)-step, step):
+                featureVectors.append(np.array([trackedObjects[j].history[i].X,trackedObjects[j].history[i].Y,trackedObjects[j].history[i].VX,trackedObjects[j].history[i].VY,trackedObjects[j].history[i+midstep].X,trackedObjects[j].history[i+midstep].Y,trackedObjects[j].history[i+step].X,trackedObjects[j].history[i+step].Y,trackedObjects[j].history[i+step].VX,trackedObjects[j].history[i+step].VY]))
+                newLabels.append(labels[j])
+                time.append(np.array([trackedObjects[j].history[i].frameID, trackedObjects[j].history[i+midstep].frameID, trackedObjects[j].history[i+step].frameID]))
+    return np.array(featureVectors), np.array(newLabels), np.array(time)
+
+def data_preprocessing_for_classifier(path2db: str, min_samples=10, max_eps=0.2, xi=0.1, min_cluster_size=10, n_jobs=18, from_half=False):
+    """Preprocess database data for classification.
+    Load, filter, run clustering on dataset then extract feature vectors from dataset.
+
+    Args:
+        path2db (str): _description_
+        min_samples (int, optional): _description_. Defaults to 10.
+        max_eps (float, optional): _description_. Defaults to 0.1.
+        xi (float, optional): _description_. Defaults to 0.15.
+        min_cluster_size (int, optional): _description_. Defaults to 10.
+        n_jobs (int, optional): _description_. Defaults to 18.
+
+    Returns:
+        List[np.ndarray]: Return X and y train and test dataset 
+    """
+    from clustering import optics_clustering_on_nx4
+    thres = 0.5
+    tracks = preprocess_database_data_multiprocessed(path2db, n_jobs=n_jobs)
+    filteredTracks = filter_out_edge_detections(tracks, threshold=thres)
+    filteredTracks = filter_tracks(filteredTracks)
+    labels = optics_clustering_on_nx4(filteredTracks, min_samples=min_samples, max_eps=max_eps, xi=xi, min_cluster_size=min_cluster_size, n_jobs=n_jobs)
+    #X, y = make_features_for_classification(filteredTracks, 6, labels)
+    if from_half:
+        X, y, time = make_features_for_classification_velocity_time_second_half(filteredTracks, 12, labels)
+    else:
+        X, y, time = make_features_for_classification_velocity_time(filteredTracks, 12, labels)
+    X = X[y > -1]
+    y = y[y > -1]
+    X_train = []
+    y_train = []
+    X_test = []
+    y_test = []
+    time_test = []
+    time_train = []
+    for i in range(len(X)):
+        if i%5==0:
+            X_test.append(X[i])
+            y_test.append(y[i])
+            time_test.append(time[i])
+        else:
+            X_train.append(X[i])
+            y_train.append(y[i])
+            time_train.append(time[i])
+    return np.array(X_train), np.array(y_train), np.array(time_train), np.array(X_test), np.array(y_test), np.array(time_test), filteredTracks
+
+def data_preprocessing_for_calibrated_classifier(path2db: str, min_samples=10, max_eps=0.2, xi=0.1, min_cluster_size=10, n_jobs=18):
+    """Preprocess database data for classification.
+    Load, filter, run clustering on dataset then extract feature vectors from dataset.
+
+    Args:
+        path2db (str): _description_
+        min_samples (int, optional): _description_. Defaults to 10.
+        max_eps (float, optional): _description_. Defaults to 0.1.
+        xi (float, optional): _description_. Defaults to 0.15.
+        min_cluster_size (int, optional): _description_. Defaults to 10.
+        n_jobs (int, optional): _description_. Defaults to 18.
+
+    Returns:
+        List[np.ndarray]: Return X and y train and test dataset 
+    """
+    from clustering import optics_clustering_on_nx4 
+    thres = 0.5
+    tracks = preprocess_database_data_multiprocessed(path2db, n_jobs=n_jobs)
+    filteredTracks = filter_out_edge_detections(tracks, threshold=thres)
+    filteredTracks = filter_tracks(filteredTracks)
+    labels = optics_clustering_on_nx4(filteredTracks, min_samples=min_samples, max_eps=max_eps, xi=xi, min_cluster_size=min_cluster_size, path2db=path2db, threshold=thres, n_jobs=n_jobs, show=True)
+    X, y = make_features_for_classification(filteredTracks, 6, labels)
+    X = X[y > -1]
+    y = y[y > -1]
+    X_train = []
+    y_train = []
+    X_calib = []
+    y_calib = []
+    X_test = []
+    y_test = []
+    first_third_limit = int(len(X) * 0.4) 
+    second_third_limit = 2*first_third_limit
+    for i in range(len(X)):
+        if i > second_third_limit-1:
+            X_test.append(X[i])
+            y_test.append(y[i])
+        elif i > first_third_limit-1 and i < second_third_limit-1: 
+            X_calib.append(X[i])
+            y_calib.append(y[i])
+        else:
+            X_train.append(X[i])
+            y_train.append(y[i])
+    return np.array(X_train), np.array(y_train), np.array(X_calib), np.array(y_calib), np.array(X_test), np.array(y_test)
+
+from classifier import BinaryClassifier
+def save_model(path2db: str, classifier_type: str, model):
+    """Save model to research_data dir.
+
+    Args:
+        path2db (str): Path to database file. 
+        classifier_type (str): Classifier name. 
+        model (Model): The model itself. 
+    """
+    if not os.path.isdir(os.path.join('research_data', path2db.split('/')[-1].split('.')[0], "models")):
+        os.mkdir(os.path.join('research_data', path2db.split('/')[-1].split('.')[0], "models"))
+    savepath = os.path.join(os.path.join('research_data', path2db.split('/')[-1].split('.')[0], "models"))
+    filename = os.path.join(savepath, f"{classifier_type}.joblib")
+    if model is not None:
+        joblib.dump(model, filename)
+    else:
+        print("Error: model is None, model was not saved.")
+
+def load_model(path2model: str) -> BinaryClassifier:
+    """Load classifier model.
+
+    Args:
+        path2model (str): Path to model. 
+
+    Returns:
+        BinaryClassifier: Trained binary classifier model. 
+    """
+    return joblib.load(path2model)
+
+def data_preprocessing_for_classifier_from_joblib_model(model: BinaryClassifier, min_samples=10, max_eps=0.2, xi=0.15, min_cluster_size=10, n_jobs=18, from_half=False):
+    """Preprocess database data for classification.
+    Load, filter, run clustering on dataset then extract feature vectors from dataset.
+
+    Args:
+        path2db (str): _description_
+        min_samples (int, optional): _description_. Defaults to 10.
+        max_eps (float, optional): _description_. Defaults to 0.1.
+        xi (float, optional): _description_. Defaults to 0.15.
+        min_cluster_size (int, optional): _description_. Defaults to 10.
+        n_jobs (int, optional): _description_. Defaults to 18.
+
+    Returns:
+        List[np.ndarray]: Return X and y train and test dataset 
+    """
+    from clustering import optics_on_featureVectors 
+    #thres = 0.5
+    featureVectors = makeFeatureVectorsNx4(model.trackData)
+    labels = optics_on_featureVectors(featureVectors, min_samples, xi, min_cluster_size, max_eps=max_eps, n_jobs=n_jobs) 
+    #X, y = make_features_for_classification(filteredTracks, 6, labels)
+    if from_half:
+        X, y, time = make_features_for_classification_velocity_time_second_half(model.trackData, 12, labels)
+    else:
+        X, y, time = make_features_for_classification_velocity_time(model.trackData, 12, labels)
+    X = X[y > -1]
+    y = y[y > -1]
+    X_train = []
+    y_train = []
+    X_test = []
+    y_test = []
+    time_test = []
+    time_train = []
+    for i in range(len(X)):
+        if i%5==0:
+            X_test.append(X[i])
+            y_test.append(y[i])
+            time_test.append(time[i])
+        else:
+            X_train.append(X[i])
+            y_train.append(y[i])
+            time_train.append(time[i])
+    return np.array(X_train), np.array(y_train), np.array(time_train), np.array(X_test), np.array(y_test), np.array(time_test)
